lite-aeolus-jerryscript: Switch worker to docker-bionic-amd64.
There was recently big hiccup with docker-xenial-amd64-13, and generally,
those workers are hogged for hours by the main Zephyr builds. OTOH, this
build takes minutes, so it makes sense to move it to more agile worker.

With that, add throttling parameters, to not hog too many
docker-bionic-amd64 builders at the same time still.

Change-Id: I8ef2619fe07c4969d61ee5a5b2b4752dac486d0c
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>