ldcg-python-tensorflow: several fixes and changes
We are building 3 different versions of TF so we need to take care of
their requirements:

 TF version | numpy  | h5py   | bazel
------------|--------|--------|-------
   1.15     | 1.18.* | 2.10   | 0.29
   2.4      | 1.19.2 | 2.10+  | 3.1.0
   HEAD     | 1.19.2 | 3.1.0+ | 3.7.4

We stick to numpy ~=1.19.2 as 1.20 dropped support for Python 3.6 so we
cannot build it under CentOS.

grpcio is set to ~=1.34.0 as this fits all three versions.

Building Python package wheels got changed so we check for newest file
after build - this way we pickup correct numpy/h5py versions when we
switch from 1.15 to 2.4 version.

Also building of Python wheels is now part of building TF to make sure
we build proper numpy/h5py versions.

Change-Id: Ia4c989dc5dcb004b85be5c8bdff90a6fafebf386