se: fix the CVE-diff report
An upstream commit in poky added LAYER name to the CVE report.
This tool compares the CVE reports, but gets confused when the
format differ between the baseline and current.

As a quick fix, just discard LAYER when it appears. Later on, when
the baseline includes LAYER, this can be revisted.

Change-Id: Ia5d1ff19efb9359f24cd7bdfd7607b1e571fa874