Enable GDB builds with all targets included and native tests
This change makes GDB build with --enable-targets=all.  This option will cause
the build system to build all possible targets, which is a good exercise.

If we attempt to run the GDB testsuite, we will run the native tests on whatever
machine we're invoking the testsuite on.

Also limit the number of parallel jobs to 1 for now.

Change-Id: I9e8106973f893732f6bd9a6c62dafe349b746ba5