lt-qcom-linux-aic100: support multiple host
Based on actual testing, we realized it's preferred to cross build the
kernel using the same target distro. This patch rework the job to use
a matrix job and build the kernel independently for each supported
distro/node. We will add centos7 shortly, once the nodes are
available.

While doing the cleanup, removed all the parameters, they are not
really needed.

Finally, these builds are intented to be used by end user, so let's
keep the last 10 artifact around, and keep all build logs.

Change-Id: I5148f86e325b8f307235645ef1e99cf7886853be
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>