lite-aeolus-micropython: Add explicit mps2_an385 build.
How it was before, is that qemu_x86 CI build also included build-and-run
for mps2_an385 (QEMU), which led to confusion. mps2_an385 in turn was
used because it's as of now the only QEMU target which is able to run
binary with embedded testsuite (~1MB in size). So, make that all explicit
by having separate mps2_an385 build target.

Change-Id: I09c42a8d8e941c4d0a3d2d3dace32d641dc918dc
Signed-off-by: Paul Sokolovsky <paul.sokolovsky@linaro.org>