tcwg_*: Prefix branch names with heads/, tags with tags/
newlib does not have release branches, so we must use a tag to point
to the latest version.

Since tcwg/round-robin.yaml.inc prefixes branch names with
refs/heads/, this makes the Jenkins git plugin to fail if the branch
name is actually a tag name.

For lack of cleaner idea, we remove "refs/" from the expansion in
tcwg/round-robin.yaml.inc, prefix branch names with "refs/" in every
instance, except where we point to the newlib tag.

Change-Id: I82140c4c587be35e2f16fc1bc89a6e0732c98e64