rpb-openembedded/publishers: fix 'latest' link when publishing
Most of our OE jobs publish to xxx/BUILD_NUMBER/{rpb,rpb-wayland} so
when we create the 'latest' link with linaro-cp, we use:
 --make-link $(dirname ${PUB_DEST})

However some jobs, such as ledge-oe-release.yaml or
lt-qcom-linux-testimages.yaml publishe without the additional 'distro'
name in PUB_DEST. In this case we end up making the 'latest' link at
the wrong place, and it does not point to the 'latest' build.

In this patch, we try to catch that special case. If PUB_DEST ends
with BUILD_NUMBER, then we use PUB_DEST for the destination 'latest'
link, which would be the case for the 2 jobs mentioned above. For all
other jobs , PUB_DEST will end with $DISTRO, not BUILD_NUMBER, so they
should be unchanged.

Note that ledge-oe-release adds a trailing '/' in PUB_DEST, so the
regexp tries to catch it too.

Change-Id: I262e61693f6de42dc9a7adee2e8f70ae34838b3e
Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>