se: simplify CVE checker
There is no need to reconstruct the CVE report using pn-buildlist.
We just need to be aware that subsequent runs of bitbake may re-build
earlier targets, resulting in changes to the CVE report.

So it suffices to just make a copy of the CVE report. We want to do this
anyway so that the filename will be fixed (no build number within), to
allow later fetching of the report via snapshots.linaro.org "latest".

Signed-off-by: Ralph Siemsen <ralph.siemsen@linaro.org>
Change-Id: I82dd8d5c909b140e7cb0c1570b4d52b6beda05be