ldcg-python-manylinux-tensorflow: more versions
On start of a job we can choose which versions we want to build:
- 1.15.x
- 2.4.x
- 2.5.x
- nightly (git HEAD)

Each of them is defined in ansible/vars/vars.yml with proper versions of
required components.

This way we can skip building old branches (1.15, 2.4) when we update
2.5.x to newer release.

It also allows to run nightly but it is not yet automated.

Change-Id: Ia8c98e3da7e95dde275b33d9276b63cb9c450256