se: add "audit=0" to lces2 kernel commandline
This suppresses printk messages from audit subsystem, which sometimes
interrupt other output, causing CI tests to fail/time-out.

The need to modify "linux_sd" macro is unfortunate, it would be nicer
to just append to bootargs, as is done for prod-image. But this needs a
larger re-factor including u-boot changes, so out of scope right now.

Change-Id: I5ba1b3bf567253c7556dee55c654751871f46e8e