tcwg-llvm-release: Update aarch64 hardware
Remove obsolete AArch64 platforms (APM, D05, AMP) and use Jade machines
instead. We set the number of used cores to 64 and the default to
Jade-03, which is a dev machine, so running a release job on 64 of its
cores every now and then shouldn't cause too much inconvenience.

Also do a drive-by fix in the name of the default git-ref from master to
main, to match upstream.

Change-Id: Ic59a9f276f5d17716dcfba64ab66ee78e705cc0a