rpb-openembedded/builders.sh: patch correct manifest
If the build is called for the GitHub PR, the builders.sh script tries
to patch repo manifest to use merge commit. However it tries to patch
.repo/manifests.xml, while we should patch .repo/manifests/default.xml.
Hardcoding manifest name is not the ideal approach, but it is definitely
better than non-patching the manifest and thus building the tip rather
than the pull request code.

Signed-off-by: Dmitry Baryshkov <dmitry.baryshkov@linaro.org>
Change-Id: I2c25981b735ed301cc271e5fa52b5578065e283b