lkft: trigger: rename the bi-weekly trigger job name
by copying the old one to the new name, and disable the old first.
After the new named jobs worked, the old jobs will be deleted.

The bi-weekly first job will be triggered on 1st, 15th every month,
and the bi-weekly-second job will be triggered on 8th, 22nd every month

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: I4f1c704e09c8ba012281af697f25a0bab904d546