lkft: hikey960:  format cache twice for docker method
on the lava side, the device is powered off just after
fastboot format command with the docker method,
and looks like the format of the cache paritition seems
not really formatted correctly, and cause the boot failures
with the encrypt commands. so as a workaround, run fastboot
reboot bootloader again to make sure the format cache
finished successfully, also tried with formatting twice,
but it seems can not fix the problem completely either.

1/10 job still failed with the problem even running the format command twice.

Like the jobs run here:
https://validation.linaro.org/scheduler/device_type/hi960-hikey?dt_length=30&dt_search=lkft-android-android-mainline-4384-boot-7668860#dt_

Signed-off-by: Yongqin Liu <yongqin.liu@linaro.org>
Change-Id: Ia4ab132b19dcad9b275841c57d218d63377d85ec