se: collect edgeagent logs
Try to understand why edgeagent intermittently fails to build. It seems
to be a problem with git submodules not being fetched. So collect a copy
of the build log files, as well as the relevant downloads directories.

This is complicated by the fact that builders.sh runs under "set -e". So
move the bitbake call into an 'if...else..fi' block, and then collect
the log files in both cases. Cannot just add "||true" since we want to
abort the rest of build in case bitbake failed.

Signed-off-by: Ralph Siemsen <ralph.siemsen@linaro.org>
Change-Id: I13c22bc88d502e6a62996ee120721b069ecb3252